refactor: 공통 fetchAPI 에 error 처리 로직 추가
